+++ src/source_unit.nim
@@ -20,13 +20,16 @@ type
 proc get_configuration*(source_filename: string): Configuration =
    # Search for a configuration file starting at the ``source_filename`` and
    # walking up to the root directory. If we find a file but fail to parse it,
-   # we fall back to default values.
+   # we fall back to default values. The ``VLS`` define is always added to the
+   # list of external defines.
    log.debug("Searching for a configuration file.")
    let filename = find_configuration_file(source_filename)
    init(result)
    if len(filename) > 0:
       try:
          result = vltoml.parse_file(filename)
+         if "VLS" notin result.defines:
+            add(result.defines, "VLS")
          log.debug("Using configuration file '$1'.", filename)
          log.debug($result)
       except ConfigurationParseError as e: